reorganize ui-node and launch
