--- devkit_ui/launch/ui.launch.py
@@ -1,14 +1,0 @@
-from launch import LaunchDescription  # type: ignore
-from launch_ros.actions import Node
-
-
-def generate_launch_description():
-    return LaunchDescription([
-        Node(
-            package='devkit_ui',
-            executable='ui_node.py',
-            name='ui_node',
-            output='screen',
-            prefix=['python3']
-        )
-    ])
