+++ devkit_ui/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/devkit_ui
+[install]
+install_scripts=$base/lib/devkit_ui
